Enhanced the documentation and the scripts

+++ scripts/setup/setup-intellix.ps1
@@ -55,32 +55,46 @@ param(
 
 if ($SqlServerInstance) {
   Write-Host "Intelligent indexing is set up on SQL Server instance $SqlServerInstance $SqlServerInstanceUser $SqlServerInstancePassword"
-  if ((-not $SqlServerInstanceUser) -or (-not $SqlServerInstancePassword)){
+  if ((-not $SqlServerInstanceUser) -or (-not $SqlServerInstancePassword)) {
     Write-Error "There are no credentials for the SQL Server instance $SqlServerInstance specified. Please use the parameters SqlServerInstanceUser and SqlServerInstancePassword to specify the credentials to access the SQL Server."
     exit 1
   }
 }
 
-$runPath = "$PSScriptRoot/run"
+Write-Verbose "Generating docker-compose file..."
 
+$runPath = "$PSScriptRoot/run"
 if (-not (Test-Path $runPath -PathType Container)) {
   mkdir $runPath -Force
 }
 
+if ($SqlServerInstance) {
+  Copy-Item -Force $PSScriptRoot/docker-compose-own-sql.template.yml $runPath/docker-compose.yml
+}
+else {
+  Copy-Item -Force $PSScriptRoot/docker-compose-sql-in-container.template.yml $runPath/docker-compose.yml
+}
+
+
 if ($IntellixDbUser -and $IntellixDbPassword) {
+  Write-Verbose "Writing connection string file..."
   if ($SqlServerInstance) {
     $server = $SqlServerInstance
-    Copy-Item -Force $PSScriptRoot/docker-compose-own-sql.template.yml $runPath/docker-compose.yml
   }
   else {
     $server = "sql\SQLEXPRESS"
-    Copy-Item -Force $PSScriptRoot/docker-compose-sql-in-container.template.yml $runPath/docker-compose.yml
   }
   $connectionString = "ConnectionStrings__IntellixDatabaseEntities=Server=$server;Database=intellixv2;user id=$intellixDbUser;password=$intellixDbPassword;Trusted_Connection=False;pooling=True;multipleactiveresultsets=True;App=Intellix"
   $connectionString | Out-File -FilePath $runPath/intellix-database.env -Encoding ASCII
+  Write-Verbose "Connection string file written"
+}
+else {
+  Write-Verbose "Skipping writing connection string file."
 }
 
 if ($IntellixAdminUser -and $IntellixAdminPassword) {
+  Write-Verbose "Writing DocuWare configuration file..."
+
   [Xml] $xml = '<?xml version="1.0"?>
   <IntellixConnectionSetup CreatedAt="2013-12-05T17:11:32.0919779+01:00" xmlns="http://dev.docuware.com/public/services/intellix">
     <ServiceUri></ServiceUri>
@@ -102,8 +116,13 @@ if ($IntellixAdminUser -and $IntellixAdminPassword) {
   $rootElement.User = $IntellixAdminUser
   $rootElement.ModelspaceName = "$($IntellixAdminUser)_Default"
   $xml.Save("$runPath/intelligent-indexing-connection.xml")
+  Write-Verbose "DocuWare configuration file written."
+}
+else {
+  Write-Verbose "Skipping DocuWare configuration file."
 }
 
+Write-Verbose "Create or update data directories..."
 $intellixDirs = @(
   'c:\ProgramData\IntellixV2'
   'c:\ProgramData\IntellixV2\SQL'
@@ -118,16 +137,22 @@ foreach ($dir in $intellixDirs) {
     mkdir $dir -Force
   }
 }
+Write-Verbose "Data directories updated."
 
-"" | Out-File -FilePath $runPath/intellix-license.env -Encoding ASCII
+if (-not (Test-Path $runPath/intellix-license.env)) {
+  "" | Out-File -FilePath $runPath/intellix-license.env -Encoding ASCII
+}
+
 if ($LicenseFile -and (Test-Path $LicenseFile)) {
   Copy-Item $LicenseFile c:\ProgramData\IntellixV2\License\license.lic
   "LicenseFileLocation=c:/license/license.lic" | Out-File -FilePath $runPath/intellix-license.env -Encoding ASCII
+  Write-Verbose "License file applied."
 }
 
 if (-not (Test-Path 'c:\ProgramData\IntellixV2\Solr\data\productionWordPairExtended' -PathType Container) ) {
   mkdir 'c:\ProgramData\IntellixV2\Solr\data' -Force
   Copy-Item  $PSScriptRoot/productionWordPairExtended 'c:\ProgramData\IntellixV2\Solr\data\' -Recurse 
+  Write-Verbose "Solr files copied."
 }
 
 $dbSetupDir = Join-Path -Path $PSScriptRoot -ChildPath database-setup
@@ -140,9 +165,22 @@ else {
 
 
 docker-compose -f $dbSetupPath build
-docker-compose -f $dbSetupPath run --rm -e intellixUserName=$IntellixAdminUser -e intellixUserPassword=$IntellixAdminPassword -e intellixDbUser=$IntellixDbUser -e intellixDbPassword=$IntellixDbPassword -e sqlServerInstance=$SqlServerInstance -e sqlServerInstanceUser=$SqlServerInstanceUser -e sqlServerInstancePassword=$SqlServerInstancePassword tools --exit-code-from tools
-docker-compose -f $dbSetupPath down
+if (!$?) {
+  Write-Error "Could not build the database setup container. Exiting..."
+  exit -1
+}
 
+docker-compose -f $dbSetupPath run --rm -e intellixUserName=$IntellixAdminUser -e intellixUserPassword=$IntellixAdminPassword -e intellixDbUser=$IntellixDbUser -e intellixDbPassword=$IntellixDbPassword -e sqlServerInstance=$SqlServerInstance -e sqlServerInstanceUser=$SqlServerInstanceUser -e sqlServerInstancePassword=$SqlServerInstancePassword tools --exit-code-from tools
+if (!$?) {
+  Write-Error "Could not run the database setup container. You should check the console output for errors. Exiting..."
+  exit -1
+}
+
+docker-compose -f $dbSetupPath down
+if (!$?) {
+  Write-Error "Could not stop the database setup container. You should try to remove the containers manually. Exiting..."
+  exit -1
+}
 
 Write-Output "Start Intelligent Indexing with 'Start-Intellix.ps1"
 Write-Output "You find the configuration file for DocuWare at '$(Join-Path $runPath 'intelligent-indexing-connection.xml')'"
